requirements.txt
# General dependencies
cchardet==2.1.6
aiodns==2.0.0
aiohttp==3.6.2
apyori==1.1.2
h5py==2.10.0
invoke==1.4.1
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.1
publicsuffixlist==0.6.13
pytest==5.4.1
PyYAML==5.3.1
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.11.0
sh==1.13.1
statsmodels==0.11.1
tensorflow==2.2.0
typing-extensions==3.7.4.1
UpSetPlot==0.4.0
jupyterlab==1.1.4

# Dependencies from VCS
git+https://github.com/jpcsmith/doceasy.git@v0.4.0#egg=doceasy
git+https://github.com/jpcsmith/wf-tools.git@v0.3.0#egg=lab
